--- a/src/qdirfiletree.h
+++ b/src/qdirfiletree.h
@@ -43,22 +43,11 @@ public:
    */
   static std::shared_ptr<const QDirFileTree> makeTree(QDir directory);
 
-  QDirFileTree(std::shared_ptr<const IFileTree> parent, QDir directory);
-
 protected:
 
+  using IFileTree::IFileTree;
 
-  virtual bool beforeReplace(IFileTree const* dstTree, FileTreeEntry const* destination, FileTreeEntry const* source) override;
-  virtual bool beforeInsert(IFileTree const* entry, FileTreeEntry const* name) override;
-  virtual bool beforeRemove(IFileTree const* entry, FileTreeEntry const* name) override;
-
-  virtual std::shared_ptr<FileTreeEntry> makeFile(std::shared_ptr<const IFileTree> parent, QString name) const override;
-  virtual std::shared_ptr<IFileTree> makeDirectory(std::shared_ptr<const IFileTree> parent, QString name) const override;
-  virtual bool doPopulate(std::shared_ptr<const IFileTree> parent, std::vector<std::shared_ptr<FileTreeEntry>>& entries) const override;
-  virtual std::shared_ptr<IFileTree> doClone() const override;
-
-  QDir qDir;
-
+  virtual bool doPopulate(std::shared_ptr<const IFileTree> parent, std::vector<std::shared_ptr<FileTreeEntry>>& entries) const = 0;
 };
 
 #endif