… src/shared/directoryentry.cpp
… src/shared/directoryentry.h
… src/version.rc