Modification to `ModInfoWithConflictInfo` to use the `ModDataChecker` feature to check mod validity:

- New `QDirFileTree` that  is a `IFileTree` populated from a `QDir`.
- Add to modify some `ModInfoXXX` class to give the game plugin to the parent `ModInfoWithConflictInfo`.

I know that the new implementation is slower than the old one. With my SkyrimSE setup with ~500 mods, I see almost no difference when looking at the code that load all mods from disk (the part of the code where this has the most impact). I don't think it will matter even for 2000 mods.

I checked the speed differences with an unrealistic setup: 2000 mods where each one has 100 folders and 100 files in the root folder (since the `QDirFileTree` does not recurse unless asked to). In this case, I go from ~900ms to 2.7s... I can go down to 1.9s by removing the `fileTime` method, but it's still slightly more than twice the old time.

**Edit:** I check the timing of `updateModInfoFromDisc` to compare, and did a bit more in-depth profiling for the 2000 mods version.

I check with another fake setup with 2000 mods, each one containing 20 folders and 20 files and the `updateModInfoFromDisc`. The timing of `updateModInfoFromDisc` goes from ~520ms to ~850ms (~720ms without the `fileTime`). This also includes all the other changes, i.e., fetching the `ModDataChecker` game-feature from the game plugin and using it, so this alone might have a slight impact.

**Edit 2:** After some modifications (not fetching `fileTime()` and sorting via `QDir::SortFlags`, I have better performances.

- For the 2000 mods containing 20 files and 20 folders, it is down to ~660ms with the new version.
- For the 2000 mods containing 100 files and 100 folders, it is down to ~1500ms.